added leetcode tasks

+++ src/main/java/org/concurrency/callable/FutureTest.java
@@ -12,18 +12,19 @@ public class FutureTest {
 
     public static void main(String[] args) {
         final ExecutorService exec = Executors.newCachedThreadPool();
-        final ArrayList<Future<Integer>> results = new ArrayList<>();
+        final ArrayList<Future<Long>> results = new ArrayList<>();
         for (int i = 10; i < 15; i++) {
             int val = i;
             results.add(exec.submit(() -> Fibonacci.fib(val)));
         }
 
         int sum = 0;
-        for (Future<Integer> result : results) {
+        for (Future<Long> result : results) {
             try {
                 sum += result.get();
             } catch (ExecutionException | InterruptedException e) {
-                System.out.print(e);            }
+                System.out.print(e);
+            }
             finally {
                 exec.shutdown();
             }
